pyproject.toml
[project]
name = "dspy-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=2.16.0",
    "dspy-ai>=2.4.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pip>=26.0.1",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
    "spacy>=3.7.0",
    "streamlit>=1.30.0",
]
